refactor: extract MileageTrackingStatsCard into its own component
The card was defined in MileageTrackingStats.tsx but rendered by both
MileageTrackingStats and MileageTrackingSummary, so the latter had to
reach into another component's file to import it.

Its BEM block moves with it — the border and radius style the card
itself, not a part of the stats block — so the class becomes
Layer__MileageTrackingStatsCard.
