+++ src/components/features/mileage/MileageTrackingStats/mileageTrackingStats.scss
@@ -20,10 +20,3 @@
 .Layer__MileageTrackingStats__Chart {
   flex: 1;
 }
-
-.Layer__MileageTrackingStats__Card {
-  box-sizing: border-box;
-
-  border-radius: var(--border-radius-2xs);
-  border: 1px solid var(--border-color);
-}
