+++ src/components/features/mileage/MileageTrackingStatsCard/mileageTrackingStatsCard.scss
@@ -1,0 +1,6 @@
+.Layer__MileageTrackingStatsCard {
+  box-sizing: border-box;
+
+  border-radius: var(--border-radius-2xs);
+  border: 1px solid var(--border-color);
+}
