+++ src/components/features/mileage/MileageTrackingSummary/MileageTrackingSummary.tsx
@@ -11,7 +11,7 @@ import { SummaryCard } from '@blocks/SummaryCard/SummaryCard'
 import { type SummaryCardInteractionProps, type SummaryCardStringOverrides, useSummaryCardSlots } from '@blocks/SummaryCard/useSummaryCardSlots'
 import { ConditionalBlock } from '@components/utility/ConditionalBlock'
 import { MileageDeductionChart } from '@features/mileage/MileageDeductionChart/MileageDeductionChart'
-import { MileageTrackingStatsCard } from '@features/mileage/MileageTrackingStats/MileageTrackingStats'
+import { MileageTrackingStatsCard } from '@features/mileage/MileageTrackingStatsCard/MileageTrackingStatsCard'
 
 import './mileageTrackingSummary.scss'
